+++ plugins/tabletools/dialogs/tableCell.js
@@ -10,6 +10,7 @@ CKEDITOR.dialog.add( 'cellProperties', function( editor ) {
 		validate = CKEDITOR.dialog.validate,
 		widthPattern = /^(\d+(?:\.\d+)?)(px|%)$/,
 		spacer = { type: 'html', html: '&nbsp;' },
+		hiddenSpacer,
 		rtl = editor.lang.dir == 'rtl',
 		colorDialog = editor.plugins.colordialog;
 
@@ -81,6 +82,7 @@ CKEDITOR.dialog.add( 'cellProperties', function( editor ) {
 							type: 'text',
 							id: 'width',
 							width: '100px',
+							requiredContent: 'td{width,height}',
 							label: langCommon.width,
 							validate: validate.number( langCell.invalidWidth ),
 
@@ -122,6 +124,7 @@ CKEDITOR.dialog.add( 'cellProperties', function( editor ) {
 						{
 							type: 'select',
 							id: 'widthType',
+							requiredContent: 'td{width,height}',
 							label: editor.lang.table.widthUnit,
 							labelStyle: 'visibility:hidden',
 							'default': 'px',
@@ -138,6 +141,7 @@ CKEDITOR.dialog.add( 'cellProperties', function( editor ) {
 						children: [ {
 							type: 'text',
 							id: 'height',
+							requiredContent: 'td{width,height}',
 							label: langCommon.height,
 							width: '100px',
 							'default': '',
@@ -149,6 +153,13 @@ CKEDITOR.dialog.add( 'cellProperties', function( editor ) {
 									labelElement = heightType.getElement(),
 									inputElement = this.getInputElement(),
 									ariaLabelledByAttr = inputElement.getAttribute( 'aria-labelledby' );
+
+								if ( this.getDialog().getContentElement( 'info', 'height' ).isVisible() ) {
+									labelElement.setHtml( '<br />' + langTable.widthPx );
+									labelElement.setStyle( 'display', 'block' );
+
+									this.getDialog().getContentElement( 'info', 'hiddenSpacer' ).getElement().setStyle( 'display', 'block' );
+								}
 
 								inputElement.setAttribute( 'aria-labelledby', [ ariaLabelledByAttr, labelElement.$.id ].join( ' ' ) );
 							},
@@ -174,10 +185,16 @@ CKEDITOR.dialog.add( 'cellProperties', function( editor ) {
 						{
 							id: 'htmlHeightType',
 							type: 'html',
-							html: '<br />' + langTable.widthPx
+							html: '',
+							style: 'display: none'
 						} ]
 					},
-					spacer,
+					hiddenSpacer = {
+						type: 'html',
+						id: 'hiddenSpacer',
+						html: '&nbsp;',
+						style: 'display: none'
+					},
 					{
 						type: 'select',
 						id: 'wordWrap',