+++ tests/plugins/tabletools/cellproperties.js
@@ -129,6 +129,39 @@
 				dialog.setValueOf( 'info', 'width', 10 );
 				dialog.setValueOf( 'info', 'widthType', 'px' );
 			} );
+		},
+
+		// https://dev.ckeditor.com/ticket/16893
+		'test allowedContent rule': function() {
+			bender.editorBot.create( {
+				name: 'required',
+				config: {
+					allowedContent: '*[lang,dir]; table tbody thead tr th td tfoot'
+				}
+			}, function( bot ) {
+				bot.setHtmlWithSelection( '<table><tr><td>Te^st</td></tr></table>' );
+				bot.dialog( 'cellProperties', function( dialog ) {
+					assert.isFalse( dialog.getContentElement( 'info', 'width' ).isVisible() );
+					assert.isFalse( dialog.getContentElement( 'info', 'height' ).isVisible() );
+					assert.isFalse( dialog.getContentElement( 'info', 'htmlHeightType' ).isVisible() );
+					assert.isFalse( dialog.getContentElement( 'info', 'hiddenSpacer' ).isVisible() );
+				} );
+			} );
+		},
+
+		// https://dev.ckeditor.com/ticket/16893
+		'test disallowedContent rule': function() {
+			bender.editorBot.create( {
+				name: 'editor'
+			}, function( bot ) {
+				bot.setHtmlWithSelection( '<table><tr><td>Te^st</td></tr></table>' );
+				bot.dialog( 'cellProperties', function( dialog ) {
+					assert.isTrue( dialog.getContentElement( 'info', 'width' ).isVisible() );
+					assert.isTrue( dialog.getContentElement( 'info', 'height' ).isVisible() );
+					assert.isTrue( dialog.getContentElement( 'info', 'htmlHeightType' ).isVisible() );
+					assert.isTrue( dialog.getContentElement( 'info', 'hiddenSpacer' ).isVisible() );
+				} );
+			} );
 		}
 	} );
 } )();